refactor: centralize shared components
