+++ frontend/src/app/app.component.ts
@@ -2,10 +2,10 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterOutlet } from '@angular/router';
-import { NavigationComponent } from './shared/components/navigation/navigation.component';
-import { ModalComponent } from './shared/components/modal/modal.component';
-import { NotificationComponent } from './shared/components/notification/notification.component';
-import { CitationFooterComponent } from "./shared/components/citation-footer/citation-footer.component";
+import { NavigationComponent } from './components/ui/navigation/navigation.component';
+import { ModalComponent } from './components/ui/modal/modal.component';
+import { NotificationComponent } from './components/ui/notification/notification.component';
+import { CitationFooterComponent } from "./components/bible/citation-footer/citation-footer.component";
 
 @Component({
   selector: 'app-root',
