+++ frontend/src/app/components/bible/citation-footer/citation-footer.component.html
