+++ frontend/src/app/components/bible/verse-range-picker/verse-picker.component.html
@@ -1,4 +1,4 @@
-<!-- frontend/src/app/shared/components/verse-range-picker/verse-picker.component.html -->
+<!-- frontend/src/app/components/bible/verse-range-picker/verse-picker.component.html -->
 <div class="verse-picker hover-reveal">
   <div class="hover-container">
     <!-- Compact Preview -->
