+++ frontend/src/app/components/bible/verse-range-picker/verse-picker.component.scss
