+++ frontend/src/app/components/memorization/memorization-modal/components/completion-stage/completion-stage.component.ts
