+++ frontend/src/app/components/memorization/memorization-modal/components/confirmation-modal/confirmation-modal.component.ts
