+++ frontend/src/app/components/memorization/memorization-modal/components/navigation-controls/navigation-controls.component.html
