+++ frontend/src/app/components/memorization/memorization-modal/components/practice-stage/practice-stage.component.ts
