+++ frontend/src/app/components/memorization/memorization-modal/components/progress-journey/progress-journey.component.html
