+++ frontend/src/app/components/memorization/memorization-modal/components/settings-menu/settings-menu.component.css
