+++ frontend/src/app/components/memorization/memorization-modal/components/setup-stage/setup-stage.component.html
