+++ frontend/src/app/components/memorization/memorization-modal/components/verse-bubbles/verse-bubbles.component.html
