+++ frontend/src/app/components/memorization/memorization-modal/memorization-modal.component.html
