+++ frontend/src/app/components/ui/modal/modal.component.scss
@@ -1,4 +1,4 @@
-// frontend/src/app/shared/components/modal/modal.component.scss
+// frontend/src/app/components/ui/modal/modal.component.scss
 .modal-overlay {
   position: fixed;
   top: 0;
