+++ frontend/src/app/components/ui/navigation/navigation.component.html
@@ -1,4 +1,4 @@
-<!-- frontend/src/app/shared/components/navigation/navigation.component.html -->
+<!-- frontend/src/app/components/ui/navigation/navigation.component.html -->
 <header class="app-header">
   <div class="header-container">
     <div class="logo-container">
