+++ frontend/src/app/components/ui/navigation/navigation.component.scss
@@ -1,4 +1,4 @@
-// frontend/src/app/shared/components/navigation/navigation.component.scss
+// frontend/src/app/components/ui/navigation/navigation.component.scss
 
 // Glassmorphism Header
 .app-header {
