+++ frontend/src/app/components/ui/notification/notification.component.scss
