+++ frontend/src/app/features/memorize/flow/components/flow-sidebar/flow-sidebar.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { VersePickerComponent, VerseSelection } from '../../../../../shared/components/verse-range-picker/verse-range-picker.component';
+import { VersePickerComponent, VerseSelection } from '../../../../../components/bible/verse-range-picker/verse-range-picker.component';
 import { FlowProgressComponent } from '../flow-progress/flow-progress.component';
 import { FlowVerse, FlowViewSettings } from '../../models/flow.models';
